YTA-2932 added Controller postfix to controllers
