+++ conf/application.conf
@@ -90,19 +90,19 @@ controllers {
     needsAuditing = false
   }
 
-  uk.gov.hmrc.epayeapi.controllers.GetTotals = {
+  uk.gov.hmrc.epayeapi.controllers.GetTotalsController = {
     needsAuth = true
     needsLogging = true
     needsAuditing = true
   }
 
-  uk.gov.hmrc.epayeapi.controllers.GetTotalsByType = {
+  uk.gov.hmrc.epayeapi.controllers.GetTotalsByTypeController = {
     needsAuth = true
     needsLogging = true
     needsAuditing = true
   }
 
-  uk.gov.hmrc.epayeapi.controllers.Documentation = {
+  uk.gov.hmrc.epayeapi.controllers.DocumentationController = {
     needsAuth = false
     needsLogging = false
     needsAuditing = false
