+++ test/uk/gov/hmrc/epayeapi/controllers/DocumentationSpec.scala
@@ -24,14 +24,14 @@ import play.api.test.Helpers._
 class DocumentationSpec extends AppSpec {
   "The API definition endpoint" should {
     "include whitelisted applications" in new App(builder.withConfig("whitelistedApplications" -> "my-application-id, my-other-application-id").build) {
-      val result = contentAsJson(inject[Documentation].definition()(FakeRequest()))
+      val result = contentAsJson(inject[DocumentationController].definition()(FakeRequest()))
 
       val lookupResult = (result \\ "whitelistedApplicationIds" ).head
       lookupResult shouldEqual JsArray(Seq(JsString("my-application-id"),
                                            JsString("my-other-application-id")))
     }
     "work without whitelisted applications" in new App(builder.build) {
-      val result = contentAsJson(inject[Documentation].definition()(FakeRequest()))
+      val result = contentAsJson(inject[DocumentationController].definition()(FakeRequest()))
 
       val lookupResult = (result \\ "whitelistedApplicationIds" ).head
       lookupResult shouldEqual JsArray()
